feat: adapt recommended test 6.2.24
