--- csaf_2_1/optionalTests.js
@@ -1,23 +1,0 @@
-export {
-  optionalTest_6_2_1,
-  optionalTest_6_2_2,
-  optionalTest_6_2_4,
-  optionalTest_6_2_5,
-  optionalTest_6_2_6,
-  optionalTest_6_2_7,
-  optionalTest_6_2_8,
-  optionalTest_6_2_9,
-  optionalTest_6_2_10,
-  optionalTest_6_2_11,
-  optionalTest_6_2_12,
-  optionalTest_6_2_13,
-  optionalTest_6_2_14,
-  optionalTest_6_2_15,
-  optionalTest_6_2_16,
-  optionalTest_6_2_17,
-  optionalTest_6_2_18,
-  optionalTest_6_2_19,
-  optionalTest_6_2_20,
-} from '../optionalTests.js'
-export { optionalTest_6_2_24 } from './optionalTests/optionalTest_6_2_24.js'
-export { optionalTest_6_2_3 } from './optionalTests/optionalTest_6_2_3.js'
