+++ csaf_2_1/recommendedTests.js
@@ -27,6 +27,7 @@ export { recommendedTest_6_2_18 } from './recommendedTests/recommendedTest_6_2_1
 export { recommendedTest_6_2_19 } from './recommendedTests/recommendedTest_6_2_19.js'
 export { recommendedTest_6_2_20 } from './recommendedTests/recommendedTest_6_2_20.js'
 export { recommendedTest_6_2_22 } from './recommendedTests/recommendedTest_6_2_22.js'
+export { recommendedTest_6_2_24 } from './recommendedTests/recommendedTest_6_2_24.js'
 export { recommendedTest_6_2_27 } from './recommendedTests/recommendedTest_6_2_27.js'
 export { recommendedTest_6_2_28 } from './recommendedTests/recommendedTest_6_2_28.js'
 export { recommendedTest_6_2_29 } from './recommendedTests/recommendedTest_6_2_29.js'
