+++ lib/shared/dateHelper.js
@@ -1,4 +1,4 @@
-import { Duration, ZonedDateTime } from '@js-joda/core'
+import { Duration, LocalDate, ZonedDateTime, ZoneId } from '@js-joda/core'
 
 /**
  * compare ZonedDateTimes from js-joda
@@ -30,4 +30,19 @@ export const compareZonedDateTimes = (a, b) => {
   } catch (e) {
     return 0
   }
+}
+
+/**
+ * Convert a local date in string representation, i.e a date accurate to the day like 'yyyy-mm-yy' to a time zoned day
+ * returns a negative number if a is less than b, positive if a is greater than b, and zero if they are equal.
+ * This function also returns 0 if one of the given values could not be parsed.
+ *
+ * @param {string} localDateString
+ * @returns the time zoned representation of the given local date
+ *
+ */
+export const localDateToTimeZonedDate = (localDateString) => {
+  const localDate = LocalDate.parse(localDateString)
+  const localDateTime = localDate.atStartOfDay()
+  return localDateTime.atZone(ZoneId.of('UTC'))
 }
